drivers/driver_ais.c: Ppacify Coverity 498043, overflow.
Fix a condition that can't happen.  Luckily (long long) can
hold both a size_t and a ssize_t, allowed arithmetic w/o
signed/unsigned issues.

+++ drivers/driver_ais.c
@@ -54,7 +54,9 @@ static void from_sixbit_untrimmed(const unsigned char *bitvec,
 // trim spaces on right end
 static void trim_spaces_on_right_end(char* to, size_t max)
 {
-    int i;
+    long long i;
+
+    // pacify Coverity 498043, overflow
     for (i = strnlen(to, max) - 1; i >= 0; i--) {
         if (' ' == to[i] ||
             '@' == to[i]) {
